chore(dockerfile): Remove bbl repo fetch
As far as I can tell, copying the bbl repo into
`var/repos/bosh-bootloader` is an outdated practice from when we wanted
to make the plan-patches and other information available to copy inside
the docker container.

Searching GitHub, I was unable to find modern examples of code that
referenced that location, so I feel good about removing this.