feat(dockerfile): Use 1.20.8-bullseye tag as base
This upgrade to bullseye, which matches Ubuntu Jammy more closely than
buster, has been a long time coming. The official Golang docker images
finally stopped supporting buster and forced our hand if we wanted to
get a later patch of Golang.

Had to change some other things in order to get this docker file to
build with the new base:
- Use the new bosh CLI dependencies
  (https://bosh.io/docs/cli-v2-install/#ubuntu).
- Upgrade to ruby 2.7.7 (what the ruby-full package installs by
  default).